+++ CHANGELOG.md
@@ -22,7 +22,8 @@ most of the commonly known power flow calculations
 -  Copy method for all `RawGridElements` and `SystemParticipants` input entities which allow an easy to use entity copy with altered field values
 -  ``distanceBetweenNodes(NodeInput nodeA, NodeInput nodeB)`` in ``GridAndGeoUtils``
 -  Additional constructors based on lists of entities in ``RawGridElements``, ``SystemParticipants`` and ``GraphicElements``
--  added ``ContainerNodeUpdateUtil`` to support updating nested nodes in ``GridContainer`` instances
+-  Added ``DistanceWeightedGraph`` + corresponding utility method to generate a graph topology whose vertices are `NodeInput` entities and its edges are weighted with the distance between the vertices in meter
+-  Added ``ContainerNodeUpdateUtil`` to support updating nested nodes in ``GridContainer`` instances
 
 ### Changed
 -  Disabled concurrent writing in `CsvFileSink.persistJointGrid()` as this caused concurrency issues